GH-29 - made FileService an interface and renamed implementation to GridFSFileService
